Listar tareas de hoy y retradas (arreglos)
