+++ spring-mvc-project/src/main/java/com/capgemini/TasksController.java
@@ -37,40 +37,11 @@ public class TasksController {
 	
 	@RequestMapping(value = "/tasks/list")
 	public String getTasks(Model model) throws SQLException {
-		
-		List<Task> tasks = tasksService.listTasks();
-		
-		
-		model.addAttribute("tasks",tasks);
-		
-		return "tasks";
-		
-	}
-	
-	@RequestMapping(value = "/tasks/todayList")
-	public String getTodayTasks(Model model) throws SQLException {
-		
-		List<Task> tasks = tasksService.listTasks();
-		
-		List<Task> todayTasks = new ArrayList<Task>();
-		
-		tasks.forEach((t)-> {
-			
-			if(t.getPlanned().before(new Date()) || t.getPlanned().getTime() == new Date().getTime())
-				todayTasks.add(t);
-			
-		});
-		
-		Collections.sort(todayTasks, new Comparator<Task>(){
-            @Override
-            public int compare(Task o1, Task o2) {
-                return o1.getCreated().compareTo(o2.getCreated());
-            }
-        });
-		
 	
 		
-		model.addAttribute("todayList",todayTasks);
+		model.addAttribute("todayList",tasksService.listTodayTasks());
+		
+		model.addAttribute("tasks",tasksService.listTasks());
 		
 		return "tasks";
